refactor: move retry utility to shared package
- Create generic retry utility in shared with logger interface
- Move retry logic to packages/shared/src/utils/retry.ts
- Update proxy retry.ts to use shared implementation with logger adapter
- Export retry utilities from shared package index
- Maintain backward compatibility for proxy service
